refactor: use tableCompanenet instead of DataTable
