+++ apps/web/src/app/[lang]/app/[type]/parties/traveller/page.tsx
@@ -1,8 +1,91 @@
 "use server";
+import type { GetApiTravellerServiceTravellersData } from "@ayasofyazilim/saas/TravellerService";
+import { $UniRefund_TravellerService_Travellers_TravellerListProfileDto } from "@ayasofyazilim/saas/TravellerService";
+import type { ColumnFilter } from "@repo/ayasofyazilim-ui/molecules/tables";
+import TableComponent from "@repo/ui/TableComponent";
 import { getResourceData } from "src/language-data/TravellerService";
-import Table from "./table";
+import { getTableData } from "../../../actions/table";
 
 export default async function Page({ params }: { params: { lang: string } }) {
   const { languageData } = await getResourceData(params.lang);
-  return <Table languageData={languageData} />;
+  type DetailedFilter = ColumnFilter & {
+    name: keyof GetApiTravellerServiceTravellersData;
+  };
+
+  const filters: DetailedFilter[] = [
+    {
+      name: "showExpired",
+      displayName: languageData["Travellers.ShowExpired"],
+      type: "boolean",
+      value: "true",
+    },
+    {
+      name: "fullName",
+      displayName: languageData["Travellers.FullName"],
+      type: "string",
+      value: "",
+    },
+    {
+      name: "fullName",
+      displayName: languageData["Travellers.FirstName"],
+      type: "string",
+      value: "",
+    },
+    {
+      name: "fullName",
+      displayName: languageData["Travellers.LastName"],
+      type: "string",
+      value: "",
+    },
+    {
+      name: "travelDocumentNumber",
+      displayName: languageData["Travellers.TravelDocumentNumber"],
+      type: "string",
+      value: "",
+    },
+    {
+      name: "username",
+      displayName: languageData["Travellers.UserName"],
+      type: "string",
+      value: "",
+    },
+    {
+      name: "phoneNumber",
+      displayName: languageData["Travellers.PhoneNumber"],
+      type: "string",
+      value: "",
+    },
+  ];
+  return (
+    <TableComponent
+      createOnNewPage
+      createOnNewPageTitle={languageData[`Travellers.New`]}
+      detailedFilter={filters}
+      fetchRequest={async (page, filter) => {
+        "use server";
+        const response = await getTableData("travellers", page, 10, filter);
+        if (response.type === "success") {
+          const data = response.data;
+          return {
+            type: "success",
+            data: { items: data.items || [], totalCount: data.totalCount || 0 },
+          };
+        }
+        return {
+          type: "success",
+          data: { items: [], totalCount: 0 },
+        };
+      }}
+      languageData={languageData}
+      tableSchema={{
+        excludeList: [
+          "id",
+          "userAccountId",
+          "residenceCountryCode2",
+          "nationalityCountryCode2",
+        ],
+        schema: $UniRefund_TravellerService_Travellers_TravellerListProfileDto,
+      }}
+    />
+  );
 }
